# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Agrivest - Agricultural Crowdfunding dApp API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  active: 'active',
  funded: 'funded',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;

export interface Campaign {
  id: number;
  title: string;
  description: string;
  farmerWallet: string;
  farmerName: string;
  goalAmount: number;
  currentAmount: number;
  deadline: string;
  category: string;
  status: CampaignStatus;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  escrowSequence?: number | null;
  /** @nullable */
  escrowTxHash?: string | null;
  investorCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface CampaignInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  farmerWallet: string;
  farmerName: string;
  /** @minimum 1 */
  goalAmount: number;
  deadline: string;
  category: string;
  imageUrl?: string;
  location?: string;
}

export interface CampaignUpdate {
  title?: string;
  description?: string;
  imageUrl?: string;
  location?: string;
}

export type InvestmentStatus = typeof InvestmentStatus[keyof typeof InvestmentStatus];


export const InvestmentStatus = {
  pending: 'pending',
  escrowed: 'escrowed',
  released: 'released',
  returned: 'returned',
} as const;

export interface Investment {
  id: number;
  campaignId: number;
  investorWallet: string;
  /** @nullable */
  investorName?: string | null;
  amount: number;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  escrowSequence?: number | null;
  status: InvestmentStatus;
  createdAt: string;
}

export interface InvestmentInput {
  investorWallet: string;
  investorName?: string;
  /** @minimum 1 */
  amount: number;
}

export interface EscrowResult {
  success: boolean;
  message: string;
  /** @nullable */
  txHash?: string | null;
}

export type PlatformStatsCategoryCountsItem = {
  category: string;
  count: number;
};

export interface PlatformStats {
  totalCampaigns: number;
  activeCampaigns: number;
  totalFunded: number;
  totalInvestors: number;
  totalXrpRaised: number;
  categoryCounts: PlatformStatsCategoryCountsItem[];
}

export type ListCampaignsParams = {
status?: ListCampaignsStatus;
category?: string;
};

export type ListCampaignsStatus = typeof ListCampaignsStatus[keyof typeof ListCampaignsStatus];


export const ListCampaignsStatus = {
  active: 'active',
  funded: 'funded',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;