# lib/api-zod/src/generated/types/campaign.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Agrivest - Agricultural Crowdfunding dApp API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignStatus } from './campaignStatus';

export interface Campaign {
  id: number;
  title: string;
  description: string;
  farmerWallet: string;
  farmerName: string;
  goalAmount: number;
  currentAmount: number;
  deadline: string;
  category: string;
  status: CampaignStatus;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  escrowSequence?: number | null;
  /** @nullable */
  escrowTxHash?: string | null;
  investorCount?: number;
  createdAt: string;
  updatedAt?: string;
}