# lib/api-zod/src/generated/types/campaignInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Agrivest - Agricultural Crowdfunding dApp API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  farmerWallet: string;
  farmerName: string;
  /** @minimum 1 */
  goalAmount: number;
  deadline: string;
  category: string;
  imageUrl?: string;
  location?: string;
}