# lib/api-zod/src/generated/types/campaignUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Agrivest - Agricultural Crowdfunding dApp API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignUpdate {
  title?: string;
  description?: string;
  imageUrl?: string;
  location?: string;
}