# lib/api-zod/src/generated/types/investment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Agrivest - Agricultural Crowdfunding dApp API
 * OpenAPI spec version: 0.1.0
 */
import type { InvestmentStatus } from './investmentStatus';

export interface Investment {
  id: number;
  campaignId: number;
  investorWallet: string;
  /** @nullable */
  investorName?: string | null;
  amount: number;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  escrowSequence?: number | null;
  status: InvestmentStatus;
  createdAt: string;
}