# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Agrivest - Agricultural Crowdfunding dApp API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformStatsCategoryCountsItem } from './platformStatsCategoryCountsItem';

export interface PlatformStats {
  totalCampaigns: number;
  activeCampaigns: number;
  totalFunded: number;
  totalInvestors: number;
  totalXrpRaised: number;
  categoryCounts: PlatformStatsCategoryCountsItem[];
}